Fixed interactive marker movement of loaded meshes, added an FSA-like controller for moving the PR2 base, some general cleanup.

+++ src/demonstration_visualizer/include/visualization_helpers.h
@@ -12,7 +12,7 @@ visualization_msgs::Marker makeMeshMarker(std::string mesh_resource,
 					  int id,
 					  const geometry_msgs::PoseStamped &pose_stamped,
 					  const geometry_msgs::Vector3 &scale,
-					  float alpha = 1.0,
+					  double alpha = 1.0,
 					  bool use_embedded_materials = true);
 
 visualization_msgs::Marker makeShapeMarker(int type,
